use u64 to represent postive BlockchainSummary values

+++ src/blocks/components.rs
@@ -726,10 +726,10 @@ pub fn BlocksSection() -> impl IntoView {
                         .get()
                         .map(|c| {
                             if c == true {
-                                summary_sig.get().blockchain_length as u64
+                                summary_sig.get().blockchain_length
                             } else {
-                                (summary_sig.get().total_num_blocks
-                                    - summary_sig.get().blockchain_length) as u64
+                                summary_sig.get().total_num_blocks
+                                    - summary_sig.get().blockchain_length
                             }
                         })
                         .or(Some(summary_sig.get().blockchain_length as u64)),
